refactored React components somewhat
