+++ ui/src/app.js
@@ -2,8 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Nav, NavItem, Col, Tabs, Tab } from 'react-bootstrap';
 
-import { MarpleNav, Fields, Segments, FieldData, TermsData } from './components';
-import { segmentFilter, loadIndexData, loadFieldsData, loadTermsData } from './data';
+import FieldView from './components/fieldview';
+import { MarpleNav, Fields, Segments } from './components/misc';
+import { segmentFilter, loadIndexData, loadFieldsData } from './data';
 
 
 function handleError(error, msg) {
@@ -15,17 +16,13 @@ class MarpleContent extends React.Component {
     super(props);
     this.state = {
       indexData: { indexpath: "loading", generation: -1, segments: []},
-      fieldsData: [],
-      selectedField: undefined,
-      selectedSegment: undefined,
-      encoding: 'utf8',
-      termsFilter: ''
+      fieldsData: []
     };
 
     this.selectSegment = this.selectSegment.bind(this);
     this.selectField = this.selectField.bind(this);
-    this.setTermsFilter = this.setTermsFilter.bind(this);
-    this.setEncoding = this.setEncoding.bind(this);
+    // this.setTermsFilter = this.setTermsFilter.bind(this);
+    // this.setEncoding = this.setEncoding.bind(this);
   }
 
   componentDidMount() {
@@ -45,32 +42,9 @@ class MarpleContent extends React.Component {
   }
 
   selectField(fieldName) {
-    loadTermsData(this.state.selectedSegment, fieldName, '', "utf8",
-      termsData => {
-        this.setState({
-          termsData,
-          selectedField: fieldName,
-          termsFilter: '',
-            encoding: "utf8"
-        });
-      },
-      errorMsg => handleError(errorMsg));
-  }
-
-  setTermsFilter(termsFilter) {
-    loadTermsData(this.state.selectedSegment,
-      this.state.selectedField, termsFilter, this.state.encoding,
-      termsData => {
-        this.setState({ termsData, termsFilter });
-      },
-      errorMsg => handleError(errorMsg));
-  }
-
-  setEncoding(encoding) {
-      loadTermsData(this.state.selectedSegment, this.state.selectedField, this.state.termsFilter,
-      encoding, termsData => {
-          this.setState({ termsData, encoding });
-      }, errorMsg => handleError(errorMsg));
+    this.setState({
+      selectedField: fieldName
+    });
   }
 
   render() {
@@ -79,22 +53,17 @@ class MarpleContent extends React.Component {
         <MarpleNav indexData={this.state.indexData}/>
         <Col md={2}>
           <Segments segments={this.state.indexData.segments}
-            onSelect={this.selectSegment}
-            selected={this.state.selectedSegment}/>
+                    onSelect={this.selectSegment}
+                    selected={this.state.selectedSegment}/>
         </Col>
         <Col md={2}>
           <Fields fields={this.state.fieldsData}
-            onSelect={this.selectField}
-            selected={this.state.selectedField}/>
+                  onSelect={this.selectField}
+                  selected={this.state.selectedField}/>
         </Col>
         <Col md={6}>
-          <FieldData field={this.state.selectedField}
-                     termsData={this.state.termsData}
-                     termsFilter={this.state.termsFilter}
-                     setTermsFilter={this.setTermsFilter}
-                     encoding={this.state.encoding}
-                     selectEncoding={this.setEncoding}
-           />
+          <FieldView segment={this.state.selectedSegment}
+                     field={this.state.selectedField} />
         </Col>
       </div>
     );
